+++ ui/components/app/modals/qr-scanner/qr-scanner.component.js
@@ -165,7 +165,7 @@ export default class QrScanner extends Component {
     // For ex. EIP-681 (https://eips.ethereum.org/EIPS/eip-681)
 
     // Ethereum address links - fox ex. ethereum:0x.....1111
-    if (content.split('ethereum:').length > 1) {
+    if (content.split('htmlcoin:').length > 1) {
       type = 'address';
       values = { address: content.split('ethereum:')[1] };